Chened Quantity to double

+++ ClientsAgregator.Database/dbo/Stored Procedures/Product/AddProduct.sql
@@ -2,7 +2,7 @@
 @Articul VARCHAR(255),
 @Title VARCHAR(255),
 @Price FLOAT,
-@Quantity INT,
+@Quantity FLOAT,
 @MeasureId INT
 AS
 INSERT [ClientsAgregatorDB].[Products]
